Fixed favicon.
There used to be a src/favicon.ico that had been created from an icon
in the Handdrawn Icon Web set. Now that a link to it has been included
in the README and the menu icons are generated from the icons in the
icon set, the favicon is now generated also.

src/images/favicon.png has been added. A new rule has been added to
the Makefile that generates public_html/favicon.ico. The old
favicon.ico has been deleted from src/.